writstturn_adapter/pyproject.toml
[project]
name = "wristturn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "ikouchiha47", email = "princeofpersiaa3.ag@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "bleak>=3.0.1",
    "pyautogui>=0.9.54",
    "aiohttp>=3.9",
    "zeroconf>=0.131",
    "pyobjc-framework-Quartz>=10.0; sys_platform == 'darwin'",
    "pyobjc-framework-Cocoa>=10.0; sys_platform == 'darwin'",
]

[project.scripts]
wristturn = "wristturn:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
